requirements.txt
absl-py==2.3.0
addict==2.4.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
dash==3.1.1
dbus-python==1.2.18
decorator==5.2.1
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==6.0.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.15
mpmath==1.3.0
munch==4.0.0
narwhals==1.45.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
plotly==6.2.0
plyfile==1.1.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyGObject==3.42.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-msssim==1.0.0
pytransform3d==3.14.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.0
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.31.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tartanair==1.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.2
tomlkit==0.12.0
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.2.0
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.20.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xformers==0.0.31
zipp==3.23.0
zstandard==0.23.0
